# apps/core-ledger-api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Core Ledger API",
    "description": "Financial Ledger REST API for institutional financial clients",
    "contact": {
      "name": "Core Ledger Team",
      "email": "support@coreledger.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7109",
      "description": "Local development (HTTPS)"
    },
    {
      "url": "http://localhost:5071",
      "description": "Local development (HTTP)"
    }
  ],
  "paths": {
...
    "/api/v1/fundos/{fundoId}/vinculos": {
      "get": {
        "tags": [
          "Vinculos"
        ],
        "operationId": "GetVinculosByFundo",
        "parameters": [
          {
            "name": "fundoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "incluirEncerrados",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CoreLedger.Application.DTOs.FundoVinculoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vinculos"
        ],
        "operationId": "CreateVinculo",
        "parameters": [
          {
            "name": "fundoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CreateFundoVinculoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.FundoVinculoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/v1/vinculos/{id}/encerrar": {
      "patch": {
        "tags": [
          "Vinculos"
        ],
        "operationId": "EncerrarVinculo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.EncerrarVinculoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoreLedger.Application.DTOs.CreateFundoVinculoDto": {
        "required": [
          "dataInicio",
          "fundoId",
          "instituicaoId",
          "tipoVinculo"
        ],
        "type": "object",
        "properties": {
          "fundoId": {
            "type": "string",
            "format": "uuid"
          },
          "instituicaoId": {
            "type": "integer",
            "format": "int32"
          },
          "tipoVinculo": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Cadastros.Enums.TipoVinculoInstitucional"
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "contratoNumero": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "observacao": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "principal": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.EncerrarVinculoDto": {
        "required": [
          "dataFim"
        ],
        "type": "object",
        "properties": {
          "dataFim": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.FundoVinculoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fundoId": {
            "type": "string",
            "format": "uuid"
          },
          "instituicaoId": {
            "type": "integer",
            "format": "int32"
          },
          "instituicaoRazaoSocial": {
            "type": "string",
            "nullable": true
          },
          "instituicaoCnpj": {
            "type": "string",
            "nullable": true
          },
          "tipoVinculo": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Cadastros.Enums.TipoVinculoInstitucional"
          },
          "tipoVinculoDescricao": {
            "type": "string",
            "nullable": true
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "dataFim": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "contratoNumero": {
            "type": "string",
            "nullable": true
          },
          "observacao": {
            "type": "string",
            "nullable": true
          },
          "principal": {
            "type": "boolean"
          },
          "estaVigente": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Domain.Cadastros.Enums.TipoVinculoInstitucional": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}